UN mission in Ethiopia misconduct allegation leak
UN personnel in Ethiopia discussing honestly an alleged orchestration by the International community to help #TPLF by all mean, which could result in the destruction of a 110M African nation--Ethiopia.

I believe this alleged recordings of this misconduct should be available to the public in light of the injustice that is being perpetrated by the international community against this poor nation of 110M. It is always sad that when an Africa and black countries get destroyed and when the population migrates to safety in Europe and America, they are again dehumanized as If it is their own sole fault. It is the alleged situations discussed in this recording that is contributing to so much and endless  suffering to poor helpless black Africans. It may not change anything but it is in African's interest to have out in the sun.